requirements.txt
# ENS Challenge 60 - Stock Auction Volume Prediction
# Dependencies

# Core Data Science
numpy>=1.21.0
pandas>=1.3.0
scipy>=1.7.0

# Machine Learning
scikit-learn>=1.0.0
lightgbm>=3.3.0
catboost>=1.0.0
xgboost>=1.5.0

# Deep Learning (optional for Phase 3)
torch>=1.10.0

# Hyperparameter Tuning
optuna>=3.0.0

# Experiment Tracking
mlflow>=2.0.0

# Visualization
matplotlib>=3.4.0
seaborn>=0.11.0
plotly>=5.0.0

# Model Interpretation
shap>=0.40.0

# Utilities
tqdm>=4.62.0
pyyaml>=6.0

# Jupyter
jupyter>=1.0.0
ipykernel>=6.0.0
ipywidgets>=7.6.0